Bug fixes + Menu bar
- Added ability to change server path
- Added About button
- Added ability to save prefs on button
- Disabled close button when server is running to stop server process hanging
- Removed time import
- Added support for filepaths with spaces in them
- Added set console text function

+++ script.py
@@ -3,10 +3,6 @@ from tkinter import filedialog
 import asyncio
 import os
 import psutil
-import time
-
-WindowWidth = 775
-WindowHeight = 347
 
 class Globals():
     ServerProc = asyncio.subprocess.Process
@@ -45,8 +41,28 @@ with open("prefs.ini",mode='r') as f:
     except: f.close()
     f.close()
 
+def set_server_path():
+    try:
+        Globals.ServerPath = filedialog.askopenfile().name
+        set_console_text("Server path set to: " + Globals.ServerPath)
+    except: 
+        set_console_text("User canceled file selection")
+    
+    Globals.LogPath = Globals.ServerPath.removesuffix("Server.exe") + "/Saved/Logs/OverlayLog.txt"
+
+def set_console_text(text:str,append = False):
+    if append:
+        Globals.ConsoleText += text
+    else:
+        Globals.ConsoleText = text
+    dpg.set_value("console_text", Globals.ConsoleText)
+
+
+WindowWidth = 775
+WindowHeight = 365
+
 dpg.create_context()
-dpg.create_viewport(title='Dread Hunger Server', width=WindowWidth, height=WindowHeight,large_icon="assets/logo.ico",small_icon="assets/logo.ico")
+dpg.create_viewport(title="Dread Hunger Server", width=WindowWidth, height=WindowHeight,large_icon="assets/logo.ico",small_icon="assets/logo.ico")
 dpg.set_exit_callback(save_prefs)
 dpg.configure_app( docking=True, docking_space=True, init_file="windows.ini")
 
@@ -57,53 +73,50 @@ with dpg.font_registry():
 def server_starter():
     asyncio.run(start_server())
 
-def kill_process():
-    for child in psutil.Process(Globals.ServerProc.pid).children(recursive=True):
-            child.kill()
+def kill_server():
+    if Globals.ServerProc is None:
+        return
     
-    Globals.ServerProc.kill()
+    if Globals.ServerProc.pid:
+        for child in psutil.Process(Globals.ServerProc.pid).children(recursive=True):
+                child.kill()
+    
+        Globals.ServerProc.kill()
 
 async def start_server():
 
-    try:
-        if Globals.ServerPath == "": Globals.ServerPath = filedialog.askopenfile().name
-    except: 
-        Globals.ConsoleText += "User canceled file selection\n"
-        dpg.set_value("console_text", Globals.ConsoleText)
-
-    Globals.LogPath = Globals.ServerPath.removesuffix("Server.exe") + "/Saved/Logs/OverlayLog.txt"
+    if Globals.ServerPath == "": 
+        set_server_path()
 
     try:
         Globals.ServerProc = await asyncio.subprocess.create_subprocess_shell(
-            f"{Globals.ServerPath} {Globals.MapCode}?maxplayers={Globals.MaxPlayers}?daysbeforeblizzard={Globals.DaysUntilBlizzard}?dayminutes={Globals.DayMinutes}?predatordamage={Globals.PredatorDamage}?coldintensity={Globals.ColdIntensity}?hungerrate={Globals.HungerRate}?coalburnrate={Globals.CoalBurnRate}?thralls={Globals.ThrallCount} -LOG=OverlayLog.txt nouniques"
+            f"{'"'}{Globals.ServerPath}{'"'} {Globals.MapCode}?maxplayers={Globals.MaxPlayers}?daysbeforeblizzard={Globals.DaysUntilBlizzard}?dayminutes={Globals.DayMinutes}?predatordamage={Globals.PredatorDamage}?coldintensity={Globals.ColdIntensity}?hungerrate={Globals.HungerRate}?coalburnrate={Globals.CoalBurnRate}?thralls={Globals.ThrallCount} -LOG=OverlayLog.txt nouniques"
         )
+
+        dpg.configure_viewport('Dread Hunger Server',disable_close=True)
     except: 
         Globals.ServerPath = ""
-        Globals.ConsoleText += "Server could not be started\n"
-        dpg.set_value("console_text", Globals.ConsoleText)
+        set_console_text("Server could not be started")
 
     while not Globals.ServerProc.returncode:
 
-        time.sleep(0.05)
+        await asyncio.sleep(0.1)
+        
         try:
-            with open(Globals.LogPath) as f: s = f.read()
-            Globals.ConsoleText = s
-            dpg.set_value("console_text", Globals.ConsoleText)
+            with open(Globals.LogPath,'r') as f: s = f.read()
+            set_console_text(s,append=True)
         except:
-            Globals.ConsoleText = "Log file not found"
-            dpg.set_value("console_text", Globals.ConsoleText)
+            set_console_text("Log file not found")
             break
 
         dpg.set_y_scroll(item="console_window",value=dpg.get_y_scroll_max(item="console_window"))
 
         if dpg.is_key_pressed(dpg.mvKey_Escape):
-            Globals.ConsoleText = "Server Interupted"
-            dpg.set_value("console_text", Globals.ConsoleText)
+            set_console_text("Server Interupted")
             break
 
-    kill_process()
-    
-    
+    dpg.configure_viewport("Dread Hunger Server",disable_close=False)
+    kill_server()
 
 def set_map(map: str):
     if map == "Approach": 
@@ -136,10 +149,17 @@ def set_cold_intensity(cold:int):
 def set_hunger(hunger:int):
     Globals.HungerRate = hunger
 
-with dpg.window(label="Console",width=WindowHeight,height=WindowWidth,no_close=True,tag="console_window"):
+with dpg.viewport_menu_bar():
+    with dpg.menu(label="File"):
+        dpg.add_menu_item(label="Set Server Path", callback=set_server_path)
+        dpg.add_menu_item(label="Save Prefs",callback=save_prefs)
+        dpg.add_menu_item(label="About",callback=lambda:dpg.configure_item("about", show=True))
+
+with dpg.window(label="Console",width=WindowHeight,height=WindowWidth,no_close=True,tag="console_window",on_close=save_prefs):
     dpg.add_text(label="",tag="console_text",wrap=WindowWidth)
 
-with dpg.window(label="Start Server",width=WindowWidth,height=WindowHeight,no_close=True, on_close=save_prefs):
+with dpg.window(label="Start Server",width=WindowWidth,height=WindowHeight,no_close=True):
+
     dpg.add_combo(items=("Approach","The Expanse","The Summit"),label="Select A Map",width=150,callback=set_map,default_value="Approach")
     dpg.add_slider_int(label="Days Until Blizzard",default_value=3,min_value=1,max_value=10,width=150, callback=set_days_until_blizz)
     dpg.add_slider_int(label="Day Minutes",default_value=8,min_value=1,max_value=50,width=150,callback=set_day_mins)
@@ -156,10 +176,9 @@ with dpg.window(label="Start Server",width=WindowWidth,height=WindowHeight,no_cl
 
     dpg.add_button(label="Start Server", callback=server_starter)
 
-if not Globals.HasOpenedBefore:
-    with dpg.window(label="Getting Started",tag="getting_started",no_docking=True,no_collapse=True,autosize=True):
-        dpg.add_text("To get started, select the settings you would like then press start server and select your DreadHungerServer exectutable.\nTo stop the server, press the esc key.", wrap=500)
-        dpg.add_button(label="Close", callback=lambda: dpg.configure_item("getting_started", show=False))
+with dpg.window(label="About",tag="about",no_docking=True,no_collapse=True,autosize=True, show=not Globals.HasOpenedBefore):
+    dpg.add_text("To get started, select the settings you would like then press start server and select your DreadHungerServer exectutable.\nTo stop the server, hold the esc key.", wrap=500)
+    dpg.add_button(label="Close", callback=lambda: dpg.configure_item("about", show=False))
 
 with dpg.theme() as global_theme:
